Rename experimental/light_curve_util/ to experimental/light_curve/
PiperOrigin-RevId: 267252085
